% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tz_coords.R
\name{convert_tz_coords}
\alias{convert_tz_coords}
\title{Create a test  function}
\usage{
convert_tz_coords(x)
}
\arguments{
\item{a}{a numeric vector}

\item{b}{a number vector}
}
\value{
a number
}
\description{
A function that adds numeric vector a and b
}
