% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tz_to_gps.R
\name{convert_tz_date}
\alias{convert_tz_date}
\title{Convert TimeZero date to POSIXct}
\usage{
convert_tz_date(x)
}
\arguments{
\item{x}{A data.frame with a 'Date' column from TimeZero.}
}
\description{
Converts integer 'Date' from a data.frame exported from a TimeZero SQLite database to a POSIXct formatted date.
}
