% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{d10_to_hex_color}
\alias{d10_to_hex_color}
\title{Convert decimal to hex color}
\usage{
d10_to_hex_color(x)
}
\arguments{
\item{x}{Numeric vector}
}
\description{
Convert a base 16 decimal color to a hex color.
}
\examples{
# Convert Globe decimal color for grey to hex (#C0C0C0)
d10_to_hex_color(12632256)
}
