% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tz_to_gps.R
\name{dd_to_dmm}
\alias{dd_to_dmm}
\title{Convert decimal degree to dmm}
\usage{
dd_to_dmm(x)
}
\arguments{
\item{x}{numeric decimal degree coordinates}
}
\description{
Convert degree decimal coordinates to degree decimal minutes (e.g. 165.5083 to 16530.5)
}
