% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_geometry_type}
\alias{detect_geometry_type}
\title{Detect geometry from an sf object}
\usage{
detect_geometry_type(x)
}
\arguments{
\item{x}{sf object}
}
\description{
Detect geometry from an sf object
}
