% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_gpx.R
\name{df_to_gpx}
\alias{df_to_gpx}
\title{Write data.frame to .gpx waypoint file}
\usage{
df_to_gpx(
  x,
  gpx_file,
  name_col,
  description_col,
  lat_col,
  lon_col,
  color_col,
  shape_col,
  gpx_format = "timezero",
  return_lines = FALSE
)
}
\arguments{
\item{x}{data.frame containing latitude, longitude, name, color, and shape columns.}

\item{gpx_file}{Output file.}

\item{name_col}{Data frame column name to use for object names.}

\item{description_col}{Description column.}

\item{lat_col}{Latitude column.}

\item{lon_col}{Longitude column.}

\item{color_col}{Color column with integer colors.}

\item{shape_col}{Shape column with integer shapes.}

\item{gpx_format}{Charactr vector indicating which marine navigation software output should be formatted for.}

\item{return_lines}{Should lines written to gpx file also be returned by the function. Used for debugging.}
}
\description{
Used to convert data.frames into a gpx file that can be read into marine navigation software.
}
