% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dmm_to_dd}
\alias{dmm_to_dd}
\title{Convert dmm format to decimal degrees}
\usage{
dmm_to_dd(x)
}
\arguments{
\item{x}{Numeric degree decimal minute coordinats}
}
\description{
Convert degree decimal minute coordinates to decimal degrees (e.g. 16530.5 to 165.5083)
}
