% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dms_string_to_dd}
\alias{dms_string_to_dd}
\title{Function to convert degree-decimal minute degree-minute-second coordinates to decimal degrees}
\usage{
dms_string_to_dd(x)
}
\arguments{
\item{x}{Character string containing latitude and longitudes separated by a comma that includes hemisphere characters (see example)}
}
\value{
Decimal degree coordinates as a matrix with with longitude in the first column and latitude in the second column (see example)
}
\description{
Function to convert degree-decimal minute degree-minute-second coordinates to decimal degrees
}
\examples{
dms_string_to_dd("72 30.500'N, 152 00.000'W")
}
