% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_cols_exist}
\alias{.check_cols_exist}
\title{Check that columns exist in a data.frame/sf}
\usage{
.check_cols_exist(x, var_cols, func_name = NULL)
}
\arguments{
\item{x}{data.frame or sf object}

\item{var_cols}{Variable columns that must be included in the data.frame}

\item{func_name}{Parent function name}
}
\description{
Internal function
}
\keyword{internal}
