% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_extra_args}
\alias{.check_extra_args}
\title{Check ellipsis for arguments and throw a warning if any are detected}
\usage{
.check_extra_args(...)
}
\arguments{
\item{...}{Passed from higher level function}
}
\description{
Check ellipsis for arguments and throw a warning if any are detected
}
\keyword{internal}
