% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_output_path}
\alias{.check_output_path}
\title{Check that filepath is valid and output directory exists}
\usage{
.check_output_path(file, ext = NULL)
}
\arguments{
\item{file}{Path to directory}

\item{ext}{Character vector of valid file extensions (e.g. c(".kml", ".gpx")).}
}
\description{
Internal function
}
\keyword{internal}
