% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_valid_geometry}
\alias{.check_valid_geometry}
\title{Check that sf geometry is valid}
\usage{
.check_valid_geometry(x, valid)
}
\arguments{
\item{x}{sf object}

\item{valid}{character vector of valid geometries}
}
\description{
Internal function
}
\keyword{internal}
