% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_globe_crabpot.R
\name{extract_globe_crabpot}
\alias{extract_globe_crabpot}
\title{Extract crab pot lines from Globe}
\usage{
extract_globe_crabpot(
  globe_file = NULL,
  raw_lines = NULL,
  raw_points = NULL,
  output_file = NULL,
  vessel_name,
  region = NULL,
  file_format = "gpkg"
)
}
\arguments{
\item{globe_file}{Path to a Globe .mdb or .accdb file, or Globe-formatted text file (.xlsx, .xls, .csv). Provide this or raw_lines and raw_points as sf objects.}

\item{raw_lines}{Converted linestrings as an sf object.}

\item{raw_points}{Converted points as an sf objects.}

\item{output_file}{Output filepath. Written to default file path if not provided.}

\item{vessel_name}{Required. Name of the vessel as a character vector.}

\item{region}{Region as a character vector.}

\item{file_format}{Geopackage "gpkg" or geodatabase "gdb"}
}
\description{
This function extracts linestrings and points from Globe formatted data and writes POINT and LINESTRING geometries to a Geopackage or Geodatabase file for subsequent review and editing. Alternatively accepts linestings and points as sf objects.
}
