% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_symbols.R
\name{find_closest_color}
\alias{find_closest_color}
\title{Find closest color}
\usage{
find_closest_color(input_color, valid_color_names)
}
\arguments{
\item{input_color}{Input color name as a character vector.}

\item{valid_color_names}{Chracter vector of valid colors.}
}
\description{
Find the color that's closest to a user-provided input color in RGB space.
}
\examples{
find_closest_color(input_color = "blue", valid_color_names = c("green", "red", "deepskyblue2"))
find_closest_color(input_color = c("magenta", "darkgreen"), valid_color_names = c("green", "red", "deepskyblue2"))
}
