% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gps_data.R
\name{get_gps_data}
\alias{get_gps_data}
\title{Get GPS data and write to rds files}
\usage{
get_gps_data(region, channel = NULL)
}
\arguments{
\item{region}{Survey region as a character vector. One of "ai", "goa", "sebs", "nbs"}

\item{channel}{RODBC connection. Function will prompt for a connection if the user is not connected}
}
\description{
Get GPS data and write to rds files
}
