% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trawl_data_for_charts.R
\name{get_trawl_data_for_charts}
\alias{get_trawl_data_for_charts}
\title{Make csv files with trawl data}
\usage{
get_trawl_data_for_charts(region, channel = NULL, species_codes = NULL)
}
\arguments{
\item{region}{Region as a character vector ("ai" or "goa")}

\item{channel}{RODBC connection. Function will prompt for a connection if the user is not connected}

\item{species_codes}{A vector of species_codes as a numeric vector. Can be used to retrieve a custom group of species.}
}
\description{
This function retrieves catch and haul data from RACEBASE for species in the GOA or AI and writes them to csv files. Files are used to produce charts for ArcMap.
}
