% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tz_data.R
\name{get_tz_data}
\alias{get_tz_data}
\title{Get vessel track data from a from TimeZero Database}
\usage{
get_tz_data(path_tzdb, start = NULL, end = NULL)
}
\arguments{
\item{path_tzdb}{A specific path to the Timezero database file (.tzdb).}

\item{start}{Optional start time as character vector POSIXct object for filtering data from the TimeZero database.}

\item{end}{Optional end time as character vector or POSIXct object for filtering ddata from the TimeZero database.}
}
\value{
A data frame containing the columns "Date", "X", and "Y" from the Timezero database.
}
\description{
This function retrieves timezone and position data from a Timezero database file.
}
\details{
The function checks if the specified file exists.
It then connects to the database, selects "Date", "X", and "Y", queries the data, and 
disconnects.
}
\examples{
\dontrun{
# Provide the path to a valid Timezero database file
path <- "path/to/timezero_database.tzdb"
data <- get_tzdata(path)
}

}
