% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_symbols.R
\name{globe_pal}
\alias{globe_pal}
\title{Globe color palette}
\usage{
globe_pal(n = NULL, values = NULL, type = "decimal", ...)
}
\arguments{
\item{n}{Number of colors to return}

\item{values}{Color values as numeric or character of colors to choose. Valid color names: "tan", "yellow", "magenta", "red", "purple", "lightgreen", "darkgreen", "cyan", "blue", "darkorange", "darkgrey", "black", "maroon", "white"}

\item{type}{Type of value to return (hex, Globe integer ("integer"), Globe decimal ("decimal"), or names).}

\item{...}{Ignored}
}
\description{
Globe color palette
}
\examples{
# List color names
globe_pal(n = Inf, type = "names")

# View colors 
show_col_nav(colors = globe_pal(n = Inf, type = "hex"))

# Globe color palette. Decimal numbers.
globe_pal(n = Inf, type = "decimal")

# Globe integer color palette
globe_pal(n = Inf, type = "integer")

# Return specific indexed color values
globe_pal(values = c(1,3,7))

# Return color values by name.
globe_pal(values = c("tan", "magenta", "darkgreen"))
}
