% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_symbols.R
\name{globe_sym_pal}
\alias{globe_sym_pal}
\title{Globe symbol palette}
\usage{
globe_sym_pal(n = NULL, values = NULL, type = "integer")
}
\arguments{
\item{n}{Optional. Number of shapes to return.}

\item{values}{Shapes to return by integer index value (0-55) or symbol name (see examples for how to show valid names names).}

\item{type}{Type of value to return (integer, names).}
}
\description{
Globe symbol palette
}
\examples{
# List symbol names
globe_sym_pal(n = Inf, type = "names")

# Globe integer values for triangle1, cirle1, and
globe_sym_pal(values = c("triangle1", "circle1", "star"), type = "integer")
}
