% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe_to_sf.R
\name{globe_to_sf}
\alias{globe_to_sf}
\title{Read Globe data from .mdb or .accdb and convert to simple features}
\usage{
globe_to_sf(
  dsn,
  tablename = NULL,
  wkt_geometry_type = NULL,
  grouping_col = NULL
)
}
\arguments{
\item{dsn}{Path to a Globe .mdb or .accdb file.}

\item{tablename}{Optional. tablename as a chracter vector (typically "marks" or "lines"). The function attempts to detect the layer type if not provided.}

\item{wkt_geometry_type}{Optional. Geometry type for output as a character vector; must be either POINT or LINESTRING. Typically "POINT" for marks or "LINESTRING" for lines.}

\item{grouping_col}{Grouping column for LINESTRINGS.}
}
\value{
A simple features object
}
\description{
Read Globe marks or lines from a Microsoft Access Database (.mdb or .accdb) and convert to an sf POINT or LINESTRING object.
}
