% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{longitude_to_utm_zone}
\alias{longitude_to_utm_zone}
\title{Longitude to UTM zone}
\usage{
longitude_to_utm_zone(x)
}
\arguments{
\item{x}{longitude}
}
\description{
Identify return the UTM zone based on longitude. UTM has 60 zones, each spanning 6 degrees of longitude.
}
