% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shapefiles_layers.R
\name{make_station_allocation}
\alias{make_station_allocation}
\title{Make station allocation shapefile and navigation plots}
\usage{
make_station_allocation(
  allocation_df,
  region,
  lon_col,
  lat_col,
  station_col,
  stratum_col,
  vessel_col,
  extra_col = NULL,
  software_format = "timezero"
)
}
\arguments{
\item{allocation_df}{Data.frame containing allocated station latitude, longitude, stratum, station ID, vessel ID, and optional additional description columns (e.g. station priority).}

\item{region}{Survey region as a character vector. One of "ai", "goa", "sebs", "nbs"}

\item{lon_col}{Name of the longitude column as a character vector. Required.}

\item{lat_col}{Name of the latitude column as a character vector. Required.}

\item{station_col}{Name of the station ID column as a character vector. Required.}

\item{stratum_col}{Name of the stratum column as a character vector. Required.}

\item{vessel_col}{Name of the vessel ID column as a character vector Required.}

\item{extra_col}{Names of columns containing data that should be included in the comments/description fields (e.g. "priority" if there's a sampling station priority field). Optional.}

\item{software_format}{Software to format output for.}

\item{survey_year}{Year for the survey. Optional.}
}
\description{
Make station allocation shapefile and navigation plots
}
