% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shapefiles_layers.R
\name{make_towpaths}
\alias{make_towpaths}
\title{Make towpath, tow start, and midpoint from GPS rds files}
\usage{
make_towpaths(
  region,
  overwrite_midpoint = FALSE,
  software_format = "timezero",
  filename_suffix = NULL
)
}
\arguments{
\item{region}{Survey region as a character vector. One of "ai", "goa", "sebs", "nbs"}

\item{overwrite_midpoint}{Should files be overwritten?}

\item{software_format}{Software format as a character vector.}

\item{filename_suffix}{Optional character vector to append to file name.}
}
\description{
Make towpath, tow start, and midpoint from GPS rds files
}
