% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tracks.R
\name{make_tracks}
\alias{make_tracks}
\title{Make track files from gps rds files}
\usage{
make_tracks(region, overwrite = FALSE)
}
\arguments{
\item{region}{Survey region as a character vector. One of "ai", "goa", "sebs", "nbs"}

\item{overwrite}{Should files be overwritten?}
}
\description{
Make track files from gps rds files
}
