% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tz_layers.R
\name{make_tz_layers}
\alias{make_tz_layers}
\title{Make kml and gpx layers for TimeZero}
\usage{
make_tz_layers(region)
}
\arguments{
\item{region}{Survey region as a character vector (passed to akfmaps::get_base_layers)}
}
\description{
Make kml and gpx layers for TimeZero
}
