% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_symbols.R
\name{navmaps_pal}
\alias{navmaps_pal}
\title{Retrieve colors for a software and file format}
\usage{
navmaps_pal(values, software_format, file_type = NULL, ...)
}
\arguments{
\item{values}{Color values as a character vector (see ?tz_pal or ?globe_pal for color options)}

\item{software_format}{Software format as a character vector.}

\item{file_type}{File type for OpenCPN or TimeZero ("kml" or "gpx")}

\item{...}{Ignored}
}
\description{
Retrieve colors for a software and file format
}
