% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_symbols.R
\name{navmaps_sym_pal}
\alias{navmaps_sym_pal}
\title{Retrieve colors for a software and file format}
\usage{
navmaps_sym_pal(values, software_format, file_type = NULL)
}
\arguments{
\item{values}{Symbol values as a character vector (see ?tz_sym_pal or ?globe_sym_pal for shape options)}

\item{software_format}{Software as a character vector.}

\item{file_type}{File type for OpenCPN or TimeZero ("kml" or "gpx")}
}
\description{
Retrieve colors for a software and file format
}
