% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_symbols.R
\name{opencpn_pal}
\alias{opencpn_pal}
\title{OpenCPN color palette}
\usage{
opencpn_pal(n = NULL, values = NULL, type = "gpx", ...)
}
\arguments{
\item{n}{Number of colors to return}

\item{values}{Color values as numeric or character of colors to choose. Valid color names: "yellow", "magenta", "red", "green", "blue", "darkorange", "black", "white"}

\item{type}{Type of value to return (hex, gpx, names).}

\item{...}{Ignored}
}
\description{
OpenCPN color palette
}
\examples{
# List color names
opencpn_pal(n = Inf, type = "names")

# View colors 
show_col_nav(colors = opencpn_pal(n = Inf, type = "hex"))

# OpenCPN named gpx color palette
opencpn_pal(n = Inf, type = "gpx")

# Return specific indexed color values
opencpn_pal(values = c(1,3,7))
}
