% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_symbols.R
\name{opencpn_sym_pal}
\alias{opencpn_sym_pal}
\title{OpenCPN symbol palette}
\usage{
opencpn_sym_pal(n = NULL, values = NULL, color = NULL, type = "names", ...)
}
\arguments{
\item{n}{Optional. Number of shapes to return.}

\item{values}{Shapes to return by gpx value (integer), kml value (URL as a character) or symbol name (see examples for how to show valid names names).}

\item{color}{Optional. Color for OpenCPN symbols to replace black. Passed to opencpn_pal().}

\item{type}{Type of value to return (gpx, kml, names).}

\item{...}{Ignored}
}
\description{
OpenCPN symbol palette
}
