% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planning_station_proximity.R
\name{planning_station_proximity}
\alias{planning_station_proximity}
\title{Calculate station proximity metrics for survey planning}
\usage{
planning_station_proximity(x, ith_station = NULL)
}
\arguments{
\item{x}{An `sf` object containing only POINT features (e.g., planning stations).}

\item{ith_station}{Should a distance be calculated for the ith nearest station to each station (e.g., 4th nearest station to each station.). If numeric, output includes distance to the ith nearest station. If NULL or 1, only returns the distance to the nearest station. (Default: `ith_station = NULL`)}
}
\value{
A list containing:
\item{nearest_stations}{A data frame with the original attributes of `x`, the index of the nearest station, 
the distance to the nearest station, and the original geometry.}
\item{n}{The number of stations.}
\item{min_nearest_m}{The minimum distance to the nearest station.}
\item{max_nearest_m}{The maximum distance to the nearest station.}
\item{sum_nearest_m}{The sum of all nearest station distances.}
\item{mean_nearest_m}{The mean distance to the nearest station.}
\item{median_nearest_m}{The median distance to the nearest station.}
}
\description{
This function calculates the proximity of points and summary statistics related to the nearest 
distances between stations. It also allows for computation of the n-th nearest station and 
distances computes various summary statistics related to the nearest distances between stations.
}
