% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_invalid_geometry}
\alias{remove_invalid_geometry}
\title{Function to remove invalid_geometries}
\usage{
remove_invalid_geometry(x)
}
\arguments{
\item{x}{an sf object}
}
\description{
Uses st_is_valid() to detect and remove invalid geometries, such as linestrings with one vertex.
}
