% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_kml_lines.R
\name{replace_kml_lines}
\alias{replace_kml_lines}
\title{Replace lines from a .kml file with a new line based on search pattern}
\usage{
replace_kml_lines(kml_path, line_pattern, replacement)
}
\arguments{
\item{kml_path}{Filepath to .kml file}

\item{line_pattern}{Pattern in the line to be replaced}

\item{replacement}{Full replacement}
}
\description{
Replace lines from a .kml file with a new line based on search pattern
}
