% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_st_write.R
\name{safe_st_write}
\alias{safe_st_write}
\title{Make shapeifles valid and write}
\usage{
safe_st_write(obj, dsn, append, ...)
}
\arguments{
\item{obj}{object of class sf or sfc}

\item{dsn}{data source name. Interpretation varies by driver: can be a filename, a folder, a database name, or a Database Connection}

\item{append}{logical; should we append to an existing layer, or replace it? if TRUE append, if FALSE replace. The default for st_write is NA which raises an error if the layer exists. The default for write_sf is FALSE, which overwrites any existing data. See also next two arguments for more control on overwrite behavior.}

\item{...}{Addition arguments passed to st_write()}
}
\description{
Wrapper around sf functions st_make_valid() and st_write(). See documentation for st_make_valid() and st_write() functions for additional arguments. sf argument descritpions are copied from the corresponding sf functions.
}
