% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_file_type}
\alias{set_file_type}
\title{File type for a software format}
\usage{
set_file_type(software_format, marks = TRUE)
}
\arguments{
\item{software_format}{Character vector of the selected software type}

\item{marks}{File format for marks (TRUE) or lines (FALSE)}
}
\description{
File type for a software format
}
