% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_globe_linestring.R
\name{sf_to_globe_linestring}
\alias{sf_to_globe_linestring}
\title{Write sf data to a csv or mdb file formatted for globe}
\usage{
sf_to_globe_linestring(x, file, color_col, time_col, extra_cols, ...)
}
\arguments{
\item{x}{sf object that contains a linestring geometry and fields with name, description, and color.}

\item{file}{Output file with a .csv or .mdb extension.}

\item{color_col}{Name of the column containing Globe color (decimal or rgb(?))}

\item{time_col}{Time column name. Optional. If not provided, writes the system time.}

\item{extra_cols}{Names of extra columns to include in output files, in the order the columns should be appended.}

\item{...}{Ignored}
}
\description{
Write sf data to a csv or mdb file formatted for globe
}
