% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_globe_track.R
\name{sf_to_globe_track}
\alias{sf_to_globe_track}
\title{Write sf data to a csv or mdb file formatted for globe}
\usage{
sf_to_globe_track(
  x,
  file,
  time_col = NULL,
  color_col = NULL,
  depth_col = NULL,
  index_col = NULL,
  ...
)
}
\arguments{
\item{x}{sf object that contains a linestring geometry and fields with name, description, and color.}

\item{file}{Output file with a .csv or .mdb extension.}

\item{time_col}{Time column name. Optional. If not provided, writes the system time.}

\item{color_col}{Name of the column containing Globe color (decimal or rgb(?))}

\item{depth_col}{Name of the depth column that contains a numeric vector of depth values}

\item{index_col}{Name of the index column that contains a numeric vector of track indices}

\item{...}{Ignored}
}
\description{
Write sf data to a csv or mdb file formatted for globe
}
