% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_gpx_route.R
\name{sf_to_gpx_route}
\alias{sf_to_gpx_route}
\title{Write sf to gpx route file}
\usage{
sf_to_gpx_route(
  x,
  file,
  name_col,
  time_col = NULL,
  color_col = NULL,
  description_col,
  ...
)
}
\arguments{
\item{x}{sf object that contains a linestring geometry and fields with name, description, and color.}

\item{file}{Output file with a .kml extension.}

\item{name_col}{Name of the column containing names.}

\item{time_col}{Time column name. Optional. If not provided, writes the system time.}

\item{color_col}{Color column name. Optional. If not provided, color is set to black.}

\item{description_col}{Description column.}

\item{...}{Ignored}
}
\description{
Write sf LINESTRING, POLYGON, MULTIPOLYGON, or MULTILINESTRING geometries to a gpx file
}
