% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_gpx_track.R
\name{sf_to_gpx_track}
\alias{sf_to_gpx_track}
\title{Write sf to gpx route file}
\usage{
sf_to_gpx_track(x, file, name_col, time_col = NULL, description_col, ...)
}
\arguments{
\item{x}{sf object that contains a linestring geometry and fields with name, description, and color.}

\item{file}{Output file with a .kml extension.}

\item{name_col}{Name of the column containing names.}

\item{time_col}{Time column name. Optional. If not provided, writes the system time.}

\item{description_col}{Description column.}

\item{...}{Ignored}
}
\description{
Write sf LINESTRING, POLYGON, MULTIPOLYGON, or MULTILINESTRING geometries to a gpx file
}
