% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_kml_points.R
\name{sf_to_kml_points}
\alias{sf_to_kml_points}
\title{Write sf data to a kml point file}
\usage{
sf_to_kml_points(
  x,
  file,
  name_col,
  description_col,
  shape_col = NULL,
  time_col = NULL,
  color_col,
  ...
)
}
\arguments{
\item{x}{sf object that contains a linestring geometry and fields with name, description, and color.}

\item{file}{Output file with a .kml extension.}

\item{name_col}{Name of the column containing names.}

\item{description_col}{Description column.}

\item{shape_col}{Name of the column containing integer shapes.}

\item{time_col}{Time column name. Optional. If not provided, writes the system time.}

\item{color_col}{Name of the column containing eight-digit hex color.}

\item{...}{Ignored}

\item{return_lines}{Should lines written to gpx file also be returned by the function. Used for debugging.}
}
\description{
Write sf POINT geometries to a kml point file
}
