% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_kml_polygon.R
\name{sf_to_kml_polygon}
\alias{sf_to_kml_polygon}
\title{Write sf data to a kml polygon file}
\usage{
sf_to_kml_polygon(
  x,
  file,
  name_col,
  description_col,
  time_col = NULL,
  color_col,
  fill_col,
  ...
)
}
\arguments{
\item{x}{sf object that contains a linestring geometry and fields with name, description, and color.}

\item{file}{Output file with a .kml extension.}

\item{name_col}{Name of the column containing names.}

\item{description_col}{Description column.}

\item{time_col}{Time column name. Optional. If not provided, writes the system time.}

\item{color_col}{Name of the column containing eight-digit hex color.}

\item{fill_col}{Name of the column containing eight-digit hex color.}

\item{...}{Ignored}
}
\description{
Write sf POLYGON or MULTIPOLYGON geometries to a kml polygon file.
}
