% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_kml_track.R
\name{sf_to_kml_track}
\alias{sf_to_kml_track}
\title{Write sf data to a kml track file for TimeZero}
\usage{
sf_to_kml_track(x, file, name_col, time_col = NULL, depth_col = NULL, ...)
}
\arguments{
\item{x}{sf object that contains a POINT or LINESTRING geometry field}

\item{file}{Output file with a .kml extension}

\item{name_col}{Name of the column containing the linestring name.}

\item{time_col}{Time column name. Optional. If not provided, writes the system time.}

\item{depth_col}{Name of the depth column that contains a numeric vector of depth values}

\item{...}{Ignored}
}
\description{
Write sf data to a kml track file for TimeZero
}
