% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_nav_file.R
\name{sf_to_nav_file}
\alias{sf_to_nav_file}
\title{Convert sf to navigation file}
\usage{
sf_to_nav_file(
  x,
  file,
  software_format,
  geometry = NULL,
  color_col = NULL,
  shape_col = NULL,
  fill_col = NULL,
  time_col = NULL,
  name_col = NULL,
  description_col = NULL,
  extra_cols = NULL,
  depth_col = NULL,
  index_col = NULL,
  tracks = FALSE,
  envir = environment()
)
}
\arguments{
\item{x}{sf object}

\item{file}{output file name}

\item{software_format}{"globe", "timezero", or "opencpn"}

\item{geometry}{Optional. Type of sf geometry as a character ("POINT", "LINESTRING",  "POLYGON", "MULTIPOINT", "MULTISTRING", "MULTIPOLYGON"). If NULL, the most common geometry type in the sf object is used.}

\item{color_col}{Column name passed to sf_to_[output_type] functions.}

\item{shape_col}{Column name passed to sf_to_[output_type] functions. Only used for POINTS and MULTIPOINT geometries.}

\item{fill_col}{Column name passed to sf_to_[output_type] functions. Only used for .kml polygons (i.e., boundaries).}

\item{time_col}{Column name passed to sf_to_[output_type] functions.}

\item{name_col}{Column name passed to sf_to_[output_type] functions. Only used for .gpx and .kml files.}

\item{description_col}{Column name passed to sf_to_[output_type] functions. Only used for .gpx and .kml files.}

\item{extra_cols}{Column names passed to sf_to_[output_type] functions. Only used for Globe .csv, .mdb, and .accdb files.}

\item{depth_col}{Column name passed to sf_to_[output_type] functions. Only used for Globe tracks created from POINT and MULTIPOINT geometries.}

\item{index_col}{Column name passed to sf_to_[output_type] functions. Only used for Globe tracks created from POINT and MULTIPOINT geometries.}

\item{tracks}{Logical. Should the output be a track file?}

\item{envir}{Only change for debugging.}
}
\description{
This function uses the sf geometry type and software_format arguments to choose an sf_to_[output_type] for writing spatial data to a file formatted for marine navigation software. For example, if the sf object passed to x contains POINT geometries and the software_format is set to "globe", the function calls sf_to_globe_points() to write data to file (.mdb, .accdb, or .csv, depending on the extension for 'file'). If the sf object passed to x contains POINT geometries and the software_format is set to "opencpn" or "timezero", the function calls sf_to_gpx_points() to write data to a .gpx file.
}
