% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_col_nav}
\alias{show_col_nav}
\title{Show marine navigation software palette colors}
\usage{
show_col_nav(
  colors = NULL,
  labels = TRUE,
  borders = NULL,
  cex_label = 1,
  ncol = NULL
)
}
\arguments{
\item{colors}{A character vector of colors.}

\item{labels}{Label each colour with its hex name?}

\item{borders}{Border colour for each tile. Default uses par("fg"). Use border = NA to omit borders.}

\item{cex_label}{Size of printed labels, as multiplier of default size.}

\item{ncol}{Number of columns. If not supplied, tries to be as square as possible.}
}
\description{
Show colors in a plot; function code and documentation based on show_col() function from the scales package
}
