% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tracks.R
\name{smooth_lat_lon_dist}
\alias{smooth_lat_lon_dist}
\title{Internal function to calculate smooth latitude and longitude and calculate distance and time elapsed}
\usage{
smooth_lat_lon_dist(df, spar = NULL)
}
\arguments{
\item{df}{data.frame of GPS data containing LONGITUDE, LATITUDE, and DATE_TIME columns}

\item{spar}{Smoothing parameter argument passed to smooth.spline()}
}
\description{
Multiple steps in this function saves computing time converting between coordinate reference systems.
}
\keyword{internal}
