% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{st_line_midpoints}
\alias{st_line_midpoints}
\title{Find the midpoint of an sf LINESTRING}
\usage{
st_line_midpoints(sf_lines = NULL)
}
\arguments{
\item{sf_lines}{sf object with LINESTRING geometries}
}
\description{
Find the midpoint of an sf LINESTRING
}
