% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_primary_centroid.R
\name{st_primary_centroid}
\alias{st_primary_centroid}
\title{Compute Centroids for Polygons and Largest Polygon in Multipolygons}
\usage{
st_primary_centroid(x)
}
\arguments{
\item{x}{An `sf` object with `POLYGON` or `MULTIPOLYGON` geometries.}
}
\value{
An `sf` object of centroid points with original attributes preserved.
}
\description{
This function calculates centroids for an `sf` object containing `POLYGON` and/or `MULTIPOLYGON` geometries. For `POLYGON`s it returns the centroid of the polygon. For `MULTIPOLYGON`s, it computes the centroid of the largest constituent polygon.
}
