% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_primary_point_on_surface.R
\name{st_primary_point_on_surface}
\alias{st_primary_point_on_surface}
\title{Compute Point on Surface for Polygons and Largest Polygon in Multipolygons}
\usage{
st_primary_point_on_surface(x)
}
\arguments{
\item{x}{An `sf` object with `POLYGON` or `MULTIPOLYGON` geometries.}
}
\value{
An `sf` object of centroid points with original attributes preserved.
}
\description{
This function calculates the point on surface for an `sf` object containing `POLYGON` and/or `MULTIPOLYGON` geometries. For `POLYGON`s it returns the point on surface of the polygon. For `MULTIPOLYGON`s, it computes the point on surface of the largest constituent polygon.
}
