% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{start_end_to_midpoint}
\alias{start_end_to_midpoint}
\title{Find midpoint between two points}
\usage{
start_end_to_midpoint(
  start_latitude,
  start_longitude,
  end_latitude,
  end_longitude,
  utm_zones = NULL
)
}
\arguments{
\item{start_latitude}{Numeric vector of latitudes in decimal degrees.}

\item{start_longitude}{Numeric vector of longitudes in decimal degrees.}

\item{end_latitude}{Numeric vector of latitudes in decimal degrees.}

\item{end_longitude}{Numeric vector of longitudes in decimal degrees.}

\item{utm_zones}{Optional numeric vector indicating which UTM zone to use for coordinates. If provided, must be either 1L or the same length as coordinate vectors. If NULL (default), the UTM zones are automatically detected from the coordinates using navmaps::longitude_to_utm_zones().}
}
\description{
Finds the midpoint between points.
}
