% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_wgs84_pseudomercator.R
\name{transform_3857_to_4326}
\alias{transform_3857_to_4326}
\title{Convert coordinates from WGS 84 Pseudo-Mercator (EPSG:3857) to WGS84 (EPSG:4326)}
\usage{
transform_3857_to_4326(x)
}
\arguments{
\item{x}{A matrix with two columns, where the first column is longitude and the second column is latitude, in EPSG:3857.}
}
\value{
A matrix with two columns, where the first column is the converted longitude and the second column is the converted latitude, in EPSG:4326.
}
\description{
This function takes a matrix of coordinates in EPSG:3857 (WGS 84 Pseudo-Mercator; used by TimeZero) and converts them to EPSG:4326 (WGS84). The input matrix is expected to have two columns, where the first column represents longitude and the second column represents latitude.
}
\examples{
# Example usage:
coordinates_3857 <- matrix(c(20037508.34, 0, 0, 20037508.34), ncol = 2, byrow = TRUE)
transformed_coordinates <- transform_3857_to_4326(coordinates_3857)

}
