% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_wgs84_pseudomercator.R
\name{transform_4326_to_3857}
\alias{transform_4326_to_3857}
\title{Convert coordinates from WGS84 (EPSG:4326) to WGS 84 Pseudo-Mercator (EPSG:3857)}
\usage{
transform_4326_to_3857(x)
}
\arguments{
\item{x}{A matrix with two columns, where the first column is longitude and the second column is latitude, in EPSG:4326.}
}
\value{
A matrix with two columns, where the first column is the converted longitude and the second column is the converted latitude, in EPSG:4326.
}
\description{
This function takes a matrix of coordinates in WGS84 (EPSG:4326) and converts them to EPSG:3857 (WGS 84 Pseudo-Mercator). The input matrix is expected to have two columns, where the first column islongitude and the second column is latitude.
}
\examples{
# Example usage:
coordinates_4326 <- matrix(c(0, 0, 180, 90), ncol = 2, byrow = TRUE)
transformed_coordinates <- transform_4326_to_3857(coordinates_4326)

}
