% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tz_pal}
\alias{tz_pal}
\title{TimeZero default color palette}
\usage{
tz_pal(n = NULL, values = NULL, software_code = TRUE)
}
\arguments{
\item{n}{Number of colors to return}

\item{values}{Color values as numeric or character of colors to choose.}

\item{software_code}{Should the full eight digit hex code (e.g., "#ffffff00") be returned? If not, returns the six digit hex equivalent (without transparency; e.g. "#ffff00").}
}
\description{
TimeZero default color palette
}
\examples{
# View colors 
show_nav_col(colors = tz_pal(n = Inf, software_code = FALSE))

# Software color palette. Eight digit hex but the first two digits represent alpha channel.
tz_pal(n = Inf, software_code = TRUE)

# Return specific indexed color values
tz_pal(values = c(1,3,7))

}
