% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_symbols.R
\name{tz_sym_pal}
\alias{tz_sym_pal}
\title{TimeZero symbol palette}
\usage{
tz_sym_pal(n = NULL, values = NULL, type = "names", ...)
}
\arguments{
\item{n}{Optional. Number of shapes to return.}

\item{values}{Shapes to return by gpx value (integer), kml value (URL as a character) or symbol name (see examples for how to show valid names names).}

\item{type}{Type of value to return (gpx, kml, names).}

\item{...}{Ignored}
}
\description{
TimeZero symbol palette
}
\examples{
# List symbol names
tz_sym_pal(n = Inf, type = "names")

# gpx symbol values for triangle1, circle1, and star
tz_sym_pal(values = c("triangle1", "circle1", "star"), type = "gpx")
 
# kml symbol values for triangle1, circle1, and star
tz_sym_pal(values = c("triangle1", "circle1", "star"), type = "kml")
}
