% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_mdb.R
\name{write_to_access}
\alias{write_to_access}
\title{Write data.frame or matrix to .mdb or .accdb}
\usage{
write_to_access(x, dsn, tablename, append = FALSE, drop_existing = TRUE)
}
\arguments{
\item{dsn}{Filepath to the .mdb or .accdb file}

\item{tablename}{character: a database table name accessible from the connected DSN. If missing, the name of dat. Passed to RODBC::sql}

\item{append}{logical. Should data be appended to an existing table? Passed to RODBC::sqlSave()}

\item{drop_existing}{Logical. Should the existing table be dropped and rewritten if it exists.}

\item{data}{data.frame to be written to the table}
}
\description{
Write data to a table in an .mdb or .accdb file. Must be run on 32-bit R, which only exists for through R 4.1!
}
