% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tz_date.R
\name{convert_tz_date}
\alias{convert_tz_date}
\title{Create a test  function}
\usage{
convert_tz_date(x)
}
\arguments{
\item{a}{a numeric vector}

\item{b}{a number vector}
}
\value{
a number
}
\description{
A function that adds numeric vector a and b
}
