% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_software}
\alias{.check_software}
\title{Check that software type is valid}
\usage{
.check_software(
  software_format,
  valid_software = c("globe", "timezero", "opencpn")
)
}
\arguments{
\item{software_format}{Character vector of the selected software type}

\item{valid_software}{Supported software options character vector.}
}
\description{
Check that software type is valid
}
