% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_table_access}
\alias{.check_table_access}
\title{Check for 32-bit driver and installation of R}
\usage{
.check_table_access(channel = NULL, schema, table_name, warn = TRUE)
}
\description{
Check for 32-bit driver and installation of R
}
