% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tzdata.R
\name{get_tzdata}
\alias{get_tzdata}
\title{Get Track Data from Timezero Database}
\usage{
get_tzdata(path_tzdb)
}
\arguments{
\item{path_tzdb}{A character string specifying the path to the Timezero database file (.tzdb).}
}
\value{
A data frame containing the columns Date, x, and y from the database.
}
\description{
This function retrieves timezone data from a Timezero database file.
}
\details{
The function checks if the specified file exists and if it has a .tzdb extension.
It then connects to the database, queries the data, and disconnects. The resulting data frame
should have at least one row; otherwise, an error is raised.
}
\examples{
\dontrun{
# Provide the path to a valid Timezero database file
path <- "path/to/timezero_database.tzdb"
data <- get_tzdata(path)
}

}
