% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{globe_pal}
\alias{globe_pal}
\title{Globe color palette}
\usage{
globe_pal(n, software_code = TRUE)
}
\arguments{
\item{n}{Number of colors to return}

\item{software_code}{Should the Globe decimal color be returned (e.g., 255 for red, 12632256 for grey) be returned? If not, returns the six digit hex equivalent (e.g. "#ff0000" for red, ).}
}
\description{
Globe color palette
}
