% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe_to_sf.R
\name{globe_to_sf}
\alias{globe_to_sf}
\title{Read Globe data from .mdb or .accdb and convert to simple features}
\usage{
globe_to_sf(
  dsn,
  tablename = NULL,
  wkt_geometry_type = NULL,
  grouping_col = NULL,
  keep_index = TRUE,
  driver = "Microsoft Access Driver (*.mdb)"
)
}
\arguments{
\item{dsn}{Path to a Globe .mdb, .accdb, .csv., .xls, or .xlsx file.}

\item{tablename}{Optional. table name as a character vector (typically "marks" or "lines"). The function attempts to detect the layer type if not provided.}

\item{wkt_geometry_type}{Optional. Geometry type for output as a character vector; must be either POINT or LINESTRING. Typically "POINT" for marks or "LINESTRING" for lines.}

\item{grouping_col}{Optional. Grouping column for LINESTRINGs.}

\item{keep_index}{Logical. Should index values from database files be retained?}

\item{driver}{odcb driver default = Microsoft Access Driver (*.mdb)}
}
\value{
A simple features object
}
\description{
Read Globe marks or lines from a Microsoft Access Database (.mdb or .accdb) and convert to an sf POINT or LINESTRING object.
}
