% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shapefiles_layers.R
\name{make_trawlable}
\alias{make_trawlable}
\title{Make trawlable/untrawlable shapefile and KML}
\usage{
make_trawlable(region, channel = NULL, software_format = "timezero")
}
\arguments{
\item{region}{Survey region as a character vector. One of "ai", "goa", "sebs", "nbs"}

\item{channel}{RODBC connection. Function will prompt for a connection if the user is not connected}

\item{software_format}{Software to format output for.}
}
\description{
Make trawlable/untrawlable shapefile and KML
}
