% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_kml_linestring.R
\name{sf_to_kml_linestring}
\alias{sf_to_kml_linestring}
\title{Write sf linestring data to a kml linestring file}
\usage{
sf_to_kml_linestring(
  x,
  kml_file,
  name_col,
  description_col,
  color_col,
  return_lines = FALSE
)
}
\arguments{
\item{x}{sf object that contains a linestring geometry and fields with name, description, and color.}

\item{kml_file}{Output file with a .kml extension.}

\item{name_col}{Name of the column containing names.}

\item{description_col}{Description column.}

\item{color_col}{Name of the column containing integer colors.}

\item{return_lines}{Should lines written to gpx file also be returned by the function. Used for debugging.}

\item{shape_col}{Name of the column containing integer shapes.}
}
\description{
Write sf linestring data to a kml linestring file
}
