% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_kml_polygon.R
\name{sf_to_kml_polygon}
\alias{sf_to_kml_polygon}
\title{Write sf data to a kml polygon file}
\usage{
sf_to_kml_polygon(
  x,
  file,
  name_col,
  description_col,
  color_col,
  fill_col,
  format = "timezero",
  return_lines = FALSE
)
}
\arguments{
\item{x}{sf object that contains a linestring geometry and fields with name, description, and color.}

\item{file}{Output file with a .kml extension.}

\item{name_col}{Name of the column containing names.}

\item{description_col}{Description column.}

\item{color_col}{Name of the column containing eight-digit hex color.}

\item{fill_col}{Name of the column containing eight-digit hex color.}

\item{return_lines}{Should lines written to gpx file also be returned by the function. Used for debugging.}

\item{shape_col}{Name of the column containing integer shapes.}
}
\description{
Write sf POLYGON or MULTIPOLYGON geometries to a kml polygon file.
}
