% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_col_nav}
\alias{show_col_nav}
\title{Show marine navigation software palette colors}
\usage{
show_col_nav(
  colors = NULL,
  custom_labels = NULL,
  labels = TRUE,
  borders = NULL,
  cex_label = 1,
  ncol = NULL,
  main = NULL,
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{colors}{A character vector of colors.}

\item{custom_labels}{Optional. Character vector to use for labels (instead of automatically using hex)}

\item{labels}{Label each color with its hex name?}

\item{borders}{Border color for each tile. Default uses par("fg"). Use border = NA to omit borders.}

\item{cex_label}{Size of printed labels, as multiplier of default size.}

\item{ncol}{Number of columns. If not supplied, tries to be as square as possible.}

\item{main}{Plot title, passed to plot()}

\item{xlab}{X-axis title, passed to plot()}

\item{ylab}{Y-axis title, passed to plot()}
}
\description{
Show colors in a plot; function code and documentation based on show_col() function from the scales package
}
