% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_towpaths.R
\name{smooth_lat_lon_dist}
\alias{smooth_lat_lon_dist}
\title{Function to calculate smooth latitude and longitude and calculate distance and time elapsed}
\usage{
smooth_lat_lon_dist(df, spar = NULL)
}
\arguments{
\item{df}{data.frame of GPS data containing LONGITUDE, LATITUDE, and DATE_TIME columns}

\item{spar}{Smoothing parameter argument passed to smooth.spline()}
}
\description{
Internal function called in make_towpaths() to smooth tow path and calculate midpoints. Multiple steps are combined in this function to save computing time converting between coordinate reference systems (geodetic WGS84 to projected UTM and back to WGS84).
}
\keyword{internal}
