% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tz_pal}
\alias{tz_pal}
\title{TimeZero default color palette}
\usage{
tz_pal(n, full_code = TRUE)
}
\arguments{
\item{n}{Number of colors to return}

\item{full_code}{Should the full eight digit hex code (e.g., "#ffffff00") be returned? If not, returns the six digit hex equivalent (without transparency; e.g. "#ffff00").}
}
\description{
TimeZero default color palette
}
